Rapidjson

+++ Engine/Src/Ancona/Core2D/Systems/Drawable/AnimatedDrawable.cpp
@@ -9,10 +9,10 @@ REGISTER_POLYMORPHIC_SERIALIZER(ild::AnimatedDrawable)
 using namespace ild;
 
 AnimatedDrawable::AnimatedDrawable(
-        const int priority,
+        const float priority,
         const std::string & key,
         float duration,
-        int priorityOffset,
+        float priorityOffset,
         sf::Vector2f anchor) :
     Drawable(
             priority,
